Collapse StatisticsController into NodeInfoController
