+++ app/assets/stylesheets/_application.scss
@@ -91,6 +91,6 @@
 @import 'highlightjs/github';
 
 /* statistics */
-@import 'new_styles/statistics';
+@import 'statistics';
 
 @import "bootstrap3-switch";
