+++ app/assets/stylesheets/statistics.scss
@@ -1,4 +1,4 @@
-.page-statistics {
+.page-node_info.action-statistics {
   h1{ text-align: center; }
 
   h3{
