+++ app/controllers/node_info_controller.rb
@@ -1,5 +1,6 @@
 class NodeInfoController < ApplicationController
   respond_to :json
+  respond_to :html, only: :statistics
 
   def jrd
     render json: NodeInfo.jrd(CGI.unescape(node_info_url("123.123").sub("123.123", "%{version}")))
@@ -11,6 +12,13 @@ class NodeInfoController < ApplicationController
       render json: document, content_type: document.content_type
     else
       head :not_found
+    end
+  end
+
+  def statistics
+    respond_to do |format|
+      format.json { render json: StatisticsPresenter.new }
+      format.all { @statistics = NodeInfoPresenter.new("1.0") }
     end
   end
 end
