+++ app/views/node_info/_statistic.haml
