+++ app/views/node_info/statistics.html.haml
@@ -1,0 +1,2 @@
+.container-fluid
+  = render "statistics"
