+++ app/views/node_info/statistics.mobile.haml
@@ -1,0 +1,1 @@
+= render "statistics"
