--- app/views/statistics/statistics.html.haml
@@ -1,2 +1,0 @@
-.container-fluid
-  = render('statistics/statistics')
