--- app/views/statistics/statistics.mobile.haml
@@ -1,1 +1,0 @@
-= render('statistics/statistics')
