+++ spec/controllers/node_info_controller_spec.rb
@@ -51,4 +51,27 @@ describe NodeInfoController do
       end
     end
   end
+
+  describe "#statistics" do
+    it "responds to format json" do
+      get :statistics, format: "json"
+      expect(response.code).to eq("200")
+    end
+
+    it "contains json" do
+      get :statistics, format: "json"
+      json = JSON.parse(response.body)
+      expect(json["name"]).to be_present
+    end
+
+    it "responds to html" do
+      get :statistics, format: "html"
+      expect(response.code).to eq("200")
+    end
+
+    it "responds to mobile" do
+      get :statistics, format: "mobile"
+      expect(response.code).to eq("200")
+    end
+  end
 end
